zeroclaw_config

Module scattered_types

Config types that were originally defined in their home modules (agent, channels, tools, trust) but are needed by the config schema. Moved here to break circular dependencies.

Structs

AutoClassifyConfig

BrowserDelegateConfig

ChannelPrecheckConfig

Channel reply-intent precheck configuration.

ClawdTalkConfig

ContextCompressionConfig

EmailConfig

EvalConfig

GmailPushConfig

HistoryPrunerConfig

ThinkingConfig

Configuration for thinking/reasoning level control.

TrustConfig

VoiceCallConfig

Enums

ThinkingLevel

How deeply the model should reason for a given message.

VoiceProvider

Which telephony provider to use.